+++ account_ebics/__manifest__.py
@@ -3,7 +3,7 @@
 
 {
     'name': 'EBICS banking protocol',
-    'version': '13.0.1.1.2',
+    'version': '13.0.1.1.3',
     'license': 'LGPL-3',
     'author': 'Noviat',
     'website': 'www.noviat.com',
